+++ CHANGELOG.md
@@ -18,6 +18,16 @@ Types of changes:
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-01-30
+
+### Added
+
+* More documentation, `pyproject.toml`
+
+### Changed
+
+* Updated compatibility to version 0.3.0 of `konsepy`
+
 ## [0.1.0]
 
 ### Added
@@ -35,6 +45,8 @@ Types of changes:
 
 * Update to meet requirements of `konsepy 0.1.0`
 
-[unreleased]: https://github.com/kpwhri/konsepy_nlp_template/compare/0.1.0...HEAD
+[unreleased]: https://github.com/kpwhri/konsepy_nlp_template/compare/v0.3.0...HEAD
+
+[0.3.0]: https://github.com/kpwhri/konsepy_nlp_template/compare/0.1.0...v0.3.0
 
 [0.1.0]: https://github.com/kpwhri/konsepy_nlp_template/releases/0.1.0